#2174 moved erg test functionality into the new Erg folder from other PR
