+++ tst/Erg/bikeergfunctions.cpp
