+++ tst/Erg/bikeergfunctions.h
